voice_navpr2/Speech-Recog/requirements.txt
langchain-community
openai>=1.59.6
python-dotenv
pyaudio
wave
langchain
playsound